+++ tests/testthat/test-freeze_dependencies.R
@@ -25,11 +25,9 @@ test_that("freeze_dependencies()", {
   file_staging <- file.path(path_staging, "packages.json")
   file_community <- file.path(path_community, "packages.json")
   file_config <- file.path(path_staging, "config.json")
-  file_staged <- file.path(path_staging, "staged.json")
   old_lines_staging <- paste(readLines(file_staging), collapse = "\n")
   old_lines_community <- paste(readLines(file_community), collapse = "\n")
   expect_false(file.exists(file_config))
-  expect_true(file.exists(file_staged))
   expect_false(old_lines_staging == old_lines_community)
   freeze_dependencies(
     path_staging = path_staging,
@@ -39,7 +37,6 @@ test_that("freeze_dependencies()", {
   expect_equal(names(config), "cran_version")
   expect_true(is.character(config$cran_version))
   expect_equal(config$cran_version, meta_snapshot()$dependency_freeze)
-  expect_false(file.exists(file_staged))
   lines_staging <- paste(readLines(file_staging), collapse = "\n")
   lines_community <- paste(readLines(file_community), collapse = "\n")
   expect_true(lines_staging == old_lines_community)